This is a POC that probably needs some refining. During TF-TRT conversion, we need to turn on different optimizers during optimization of the main graph and `TRTEngine` functions. Since other functions are inlined, it is effectively equivalent to using different optimizers for the function library.

### How it works

- Add fields to `RewriterConfig` to specify an alternate list of optimizers and custom optimizers, along with a boolean to toggle the use of these alternate lists. The default behavior doesn't change (i.e same optimizers for the main graph and function library)
- Add an optional parameter to `MetaOptimizer::OptimizeGraph()` and helper functions to know whether the main graph or function library is being optimized, and set this parameter when optimizing the function library in `MetaOptimizer::OptimizeConsumeItem()`.
- When optimizing the function library, get the optimizer list from these new fields if the boolean is true.

### Example of TRT conversion logs (BERT Base)

```
Optimization results for grappler item: tf_graph
  constant_folding: Graph size after: 2884 nodes (-409), 3897 edges (-410), time = 2186.9751ms.
  layout: Graph size after: 2884 nodes (0), 3897 edges (0), time = 1524.755ms.
  constant_folding: Graph size after: 2884 nodes (0), 3897 edges (0), time = 447.266ms.
  TensorRTOptimizer: Graph size after: 521 nodes (-2363), 664 edges (-3233), time = 4766.79492ms.
  constant_folding: Graph size after: 229 nodes (-292), 445 edges (-219), time = 247.384ms.
Optimization results for grappler item: TRTEngineOp_0_2_native_segment
  constant_folding: Graph size after: 8 nodes (0), 7 edges (0), time = 1.111ms.
Optimization results for grappler item: TRTEngineOp_0_1_native_segment
  constant_folding: Graph size after: 36 nodes (0), 37 edges (0), time = 45.684ms.
Optimization results for grappler item: TRTEngineOp_0_0_native_segment
  constant_folding: Graph size after: 2689 nodes (0), 3132 edges (0), time = 455.876ms.
```

### TODO

- [ ] Gather feedback
- [ ] Add tests
- [ ] Change defaults in `trt_convert.py`